fixed logged criteria

+++ src/books/views.py
@@ -62,15 +62,20 @@ def ViewBook(request, bid):
             # Book is not among the last 20 books viewed by the user.
             ViewedBook = View(user=request.user, bid=bid, lastviewed=timezone.now())
             ViewedBook.save()
-    if request.method == 'POST':
-        book = Save.objects.filter(Q(user=request.user), Q(bid=bid)).first()
-        if book:
-            book.delete()
-            return redirect('books:ViewBook', bid=bid)
-        else:
-            book = Save(user=request.user, bid=bid)
-            book.save()
-            return redirect('books:ViewBook', bid=bid)
+    if request.user.is_authenticated():
+        if request.method == 'POST':
+            book = Save.objects.filter(Q(user=request.user), Q(bid=bid)).first()
+            if book:
+                book.delete()
+                return redirect('books:ViewBook', bid=bid)
+            else:
+                book = Save(user=request.user, bid=bid)
+                book.save()
+                return redirect('books:ViewBook', bid=bid)
+    else: 
+        messages.error(
+            request, f"Please login to save book"
+        )
     if isSaved:
         return render(request,'booksaved.html', {'detail': detail, 'RatedBook': RatedBook})
     else:
